COREUI-289 common: add pipeline

+++ scripts/stop.sh
@@ -2,4 +2,8 @@
 
 read port <<< $(lsof -i :3031 | grep -v 'PID'| awk '{print $2}')
 echo $port
-kill $port
+
+if [ "$port" ]
+  then
+    kill $port
+fi
